+++ src/game/objects/inanimates.js
@@ -16,129 +16,6 @@ import Inanimate from './Inanimate.js';
 import log from '../../lib/log.js';
 
 /**
- * Comparator function for comparing two inanimate objects
- *
- * @param {Object} _left  - Inanimate object #1
- * @param {Object} _right - Inanimate object #2
- *
- * @return {Boolean} True if equal, false otherwise
- */
-function inanimateNameComparitor(_left, _right) {
-  const left = _left.toLowerCase();
-  const right = _right.toLowerCase();
-
-  return left === right;
-}
-
-/**
- * A container that holds inanimate objects
- */
-class InanimateContainer {
-
-  /**
-   * Create a new inanimate container
-   */
-  constructor() {
-    this.inanimates = [];
-  }
-
-  /**
-   * The number of items in the container
-   *
-   * @return {Number}
-   */
-  get length() {
-    return this.inanimates.length;
-  }
-
-  /**
-   * Return the underlying Array containing all the items
-   *
-   * @return {Array}
-   */
-  get all() {
-    return this.inanimates;
-  }
-
-  /**
-   * Find and remove an item by name
-   *
-   * This uses case insensitive searching with '.' syntax to disambiguate
-   * multiple matches
-   *
-   * @param {String} name - The name of the item to remove
-   *
-   * @return {Object} - The removed item or null
-   */
-  findAndRemoveItem(name) {
-    const item = this.findItem(name);
-    if (!item) {
-      return null;
-    }
-
-    return this.removeItem(item);
-  }
-
-  /**
-   * Find an item
-   * This uses case insensitive searching with '.' syntax to disambiguate
-   * multiple matches
-   *
-   * @param {String} name - The name of the item
-   *
-   * @return {Object} - the item or null
-   */
-  findItem(name) {
-    let which = 0;
-    if (name.includes('.')) {
-      const tokens = name.split('.');
-      if (tokens.length !== 2) {
-        return null;
-      }
-      which = parseInt(tokens[0]);
-      name = tokens[1];
-      if (isNaN(which) || name.length === 0) {
-        return null;
-      }
-    }
-
-    const items = this.inanimates.filter(item => inanimateNameComparitor(item.name, name));
-    if (items.length === 0 || which >= items.length) {
-      return null;
-    }
-    const item = items[which];
-
-    return item;
-  }
-
-  /**
-   * Add an item
-   *
-   * @param {Object} item - The inanimate object to add
-   */
-  addItem(item) {
-    this.inanimates.push(item);
-  }
-
-  /**
-   * Removes an item from the container
-   *
-   * @param {Object} item - The inanimate object to remove
-   *
-   * @returns {Object} item or null
-   */
-  removeItem(item) {
-    const index = this.inanimates.indexOf(item);
-    if (index > -1) {
-      this.inanimates.splice(index, 1);
-      return item;
-    }
-
-    return null;
-  }
-}
-
-/**
  * Loads an inanimate object and its model and returns an instantiated object
  *
  * @param {Object} param
@@ -181,7 +58,5 @@ async function loadInanimate(param) {
 
 
 export {
-  inanimateNameComparitor,
-  InanimateContainer,
   loadInanimate,
 };
